+++ frontend/src/components/dashboard/MuseumGallery.js
@@ -11,7 +11,7 @@ import Select from '@material-ui/core/Select';
 import MenuItem from '@material-ui/core/MenuItem';
 import {makeStyles} from '@material-ui/core/styles';
 import Pagination from '@material-ui/lab/Pagination';
-import {fetchMuseumArtworks} from '../../redux/museumArtworkActions';
+import {fetchMuseumArtworks, getRandomArtworkId} from '../../redux/museumArtworkActions';
 import {useDispatch, useSelector} from 'react-redux';
 import Typography from '@material-ui/core/Typography';
 import LinearProgress from '@material-ui/core/LinearProgress';
@@ -91,22 +91,22 @@ export default function MuseumGallery() {
 
   const [sortBy, setSortBy] = React.useState('relevance');
   const handleChangeSortBy = (event, value) => {
-    setNewQuery(true);
     setSortBy(value);
     handleCloseSortByMenu();
+    setNewQuery(true);
   };
   const [museumPage, setMuseumPage] = useState(1);
   const handleChangePage = (event, value) => {
-    setNewQuery(true);
     setMuseumPage(value);
+    setNewQuery(true);
   };
   const [searchQuery, setSearchQuery] = React.useState('');
   const handleChangeSearch = () => {
-    setNewQuery(true);
     setSearchQuery(document.getElementById('search-textfield').value);
     setMuseumPage(1);
     document.getElementById('search-bar').style.marginBottom = '0px';
     document.getElementById('filter-drawer').style.display = 'block';
+    setNewQuery(true);
   };
   const clearSearchQuery = () => {
     document.getElementById('filter-drawer').style.display = 'none';
@@ -141,14 +141,26 @@ export default function MuseumGallery() {
   const artworks = Array.from(artworksMap);
   const dispatch = useDispatch();
   const LIMIT = 9;
+  const [firstPgLoad, setFirstPgLoad] = React.useState(true);
+  const FIRST_PAGE = 1;
+  const EMPTY_QUERY = '';
   useEffect(() => {
+    if (firstPgLoad) {
+      // resets artworks when users hit back button
+      dispatch(fetchMuseumArtworks(FIRST_PAGE, LIMIT, EMPTY_QUERY))
+          // .then(dispatch(getRandomArtworkId(
+          //     LIMIT, searchQuery, sortBy, searchField, numOfResults)))
+          .then(setFirstPgLoad(false));
+    }
     if (newQuery) {
       dispatch(fetchMuseumArtworks(
           museumPage, LIMIT, searchQuery, sortBy, searchField))
+          // .then(dispatch(getRandomArtworkId(
+          //     LIMIT, searchQuery, sortBy, searchField, numOfResults)))
           .then(setNewQuery(false));
     }
   }, [dispatch, newQuery, museumPage, searchQuery,
-    sortBy, searchField, newQuery]);
+    sortBy, searchField, firstPgLoad]);
 
   let paginationNeeded = true;
   let results;
@@ -213,12 +225,7 @@ export default function MuseumGallery() {
             </IconButton>
           </div>
         </Container>
-        <SurpriseMeButton
-          numOfResults={numOfResults}
-          searchQuery={searchQuery}
-          sortBy={sortBy}
-          searchField={searchField}
-        />
+        <SurpriseMeButton/>
       </Container>
       <Container id="filter-drawer" className={classes.filterDrawer}>
         <Button
